requirements.txt
streamlit
pandas
numpy
plotly
torch
torchaudio
transformers
accelerate
sentencepiece
openai-whisper
pydub